Add `playlist_from_upload` function
Why?
Users may want to preview the tracks uploaded by another user so they
can choose which, if any, they want to download.

What?
Adds a `playlist_from_upload` function to the
`spotify.spotifyplaylist_builder` module.

This function can read the output of the Discord webhook from the system
clipboard (or from a file) and search Spotify for those tracks creating
a playlist called "<USER> Uploads" where <USER> is a top-level user
directory sampled from the webhook output.

A new configuration option `PLAYLIST_FROM_UPLOAD` has been added which
accepts a boolean or a file path. Discord webhook output must be in the
system clipboard if set to `true`. If set to a file path, that file
obviously must exist and contain the output of the Discord webhook.