Download Beatcloud tracks from Spotify a playlist
Why?
Users may not want to download the entire upload of another user.

What?
Given a Spotify playlist with a subset of tracks from a user's upload,
perhaps generated with `PLAYLIST_FROM_UPLOAD`, compare those tracks with
the Beatcloud files and add the Beatcloud paths of matching tracks to
`DOWNLOAD_INCLUDE_DIRS`.