Refactor spotify and utils packages (#75)
* Refactor packages

Why?
The membership of functions to their parent subpackages was misplaced
in many places. Also, the primary modules accessed as features from the
top-level should be stripped down to the core functions with all other functions
moved into helper modules. Reorganizing things to minimize the amount of
cross-package dependency, and reducing the amount of code in the
feature-oriented modules, makes the code easier to read and maintain.

What?
Isolate code between the different packages as much as possible. Move all
non-core functions in the feature-oriented modules into helper modules.