Configuration refactor
Why?
`configs` should be made a first-class package with config building,
validation, and CLI argument parsing included there instead of in `utils`.

The inter-package dependencies should be minimized to improve readability and
maintainability of the code base; this also improves usability for end-users as
it improves clarity with respect to which config options / CLI args map to
which features of the library.

Configuration files should be in YAML format rather than JSON because YAML
requires fewer characters and is easy to read and write. This reduces the
surface area for user-error when editing the configuration files.

Pydantic models should be used to perform validation at the point of loading
automatically and encourage separation of concern between the different
packages' configuration options.

What?
Make `configs` a first-class Python package. Move the config-related utilities
from the `utils` folder to `configs.helpers`.

Move `COPY_TRACKS` feature from `utils` to `rekordbox`. Rename configuration
options for more clarity and consistency. Replace underscore characters in CLI
args with dashes to make them easier to type.

Replaced all JSON configuration files with equivalent YAML files. `config.yaml`
options are now grouped by their respective package. Options that either do not
apply to a particular package or else apply to more than one package are
assigned to the `configs` package.

Each package's configuration options are loaded into Pydantic models that
perform validation. All exception handling that can be validated at the point
of loading (which is almost all of it) is not being performed during Pydantic
model loading instead of in the `build_config` function or at the points of
execution.