Consolidate configs
Why?
Passing package-specific configurations through the code causes issues
when a function depends on attributes of a different package's config. The
two solutions for this are to (1) add those config options to the
`BaseConfig` or (2) pass multiple configs to the functions where needed.
Neither of these solutions are good though. The first leads to confusion
for users since configuration options that seem to belong to a specific
subpackage need to be declared in the "configs" section of
`config.yaml`. The second solution leads to confusion for developers
because it introduces inconsistencies in the function interfaces.

What?
Continue using package-level configuration objects for the purposes of
validation, but merge all the config objects into one `BaseConfig` which is
returned from the `build_config` function.