Releases/2.4.0 (#84)
* [FEATURE] Add script for moving Comments field "My Tag" data to Genre field

Why?
Some users may have captured Genre information in the "My Tag" system
which adds clutter to their tracks' Comments fields.

What?
Given a list of Genre tags that are populated by the My Tags system and
written to the Comments field, write those tags to the Genre field.

Note: these tags aren't removed from the Comments field since that can
easily be done within the Rekordbox software itself.

* [FEATURE] Create Spotify playlist from Discord webhook output

Why?
Users may want to preview the tracks uploaded by another user so they
can choose which, if any, they want to download.

What?
Adds a `playlist_from_upload` function to the
`spotify.spotify_playlist_builder` module.

This function can read the output of the Discord webhook from the system
clipboard (or from a file) and search Spotify for those tracks creating
a playlist called "<USER> Uploads" where <USER> is a top-level user
directory sampled from the webhook output.

A new configuration option `PLAYLIST_FROM_UPLOAD` has been added which
accepts a boolean or a file path. Discord webhook output must be in the
system clipboard if set to `true`. If set to a file path, that file
obviously must exist and contain the output of the Discord webhook.

* [FEATURE] Download Beatcloud tracks from Spotify a playlist

Why?
Users may not want to download the entire upload of another user.

What?
Given a Spotify playlist with a subset of tracks from a user's upload,
perhaps generated with `PLAYLIST_FROM_UPLOAD`, compare those tracks
with Beatcloud files and add the paths of matching tracks to
`DOWNLOAD_INCLUDE_DIRS`.

* [ENHANCEMENT] Refactor packages and configuration

Why?
The membership of functions to their parent subpackages was misplaced
in many places. Also, the primary modules accessed as features from the
top-level should be stripped down to the core functions with all other functions
moved into helper modules. Reorganizing things to minimize the amount of
cross-package dependency, and reducing the amount of code in the
feature-oriented modules, makes the code easier to read and maintain.

`configs` should be made a first-class package with config building,
validation, and CLI argument parsing included instead of in living in `utils`.

The inter-package dependencies should be minimized to improve readability and
maintainability of the code base; this also improves usability for end-users as
it improves clarity with respect to which config options / CLI args map to
which features of the library.

Configuration files should be in YAML format rather than JSON because YAML
requires fewer characters and is easy to read and write. This reduces the
surface area for user-error when editing the configuration files.

Pydantic models should be used to perform validation at the point of loading
automatically and encourage separation of concern between the different
packages' configuration options.

Main function needs unit testing.

What?
Isolate code between the different packages as much as possible. Move all
non-core functions in the feature-oriented modules into helper modules.

Make `configs` a first-class Python package. Move the config-related utilities
from the `utils` folder to `configs.helpers`.

Move `COPY_TRACKS` feature from `utils` to `rekordbox`. Rename configuration
options for more clarity and consistency. Replace underscore characters in CLI
args with dashes to make them easier to type.

Replaced all JSON configuration files with equivalent YAML files. `config.yaml`
options are now grouped by their respective package. Options that either do not
apply to a particular package or else apply to more than one package are
assigned to the `configs` package.

Each package's configuration options are loaded into Pydantic models that
perform validation. All exception handling that can be validated at the point
of loading (which is almost all of it) is not being performed during Pydantic
model loading instead of in the `build_config` function or at the points of
execution.

Refactor the main function out of `dj_tools.py` and into a separate module.

* [BUGFIX] Fix python-Levenshtein package version

Why?
A pip bug is causing a name conflict when importing python-Levenshtein.

What?
Fix python-Levenshtein version at 0.12.2.

More info:
https://github.com/maxbachmann/python-Levenshtein/issues/1