[FIX] Serialize `USB_PATH` as a string.
Why?
`USB_PATH` was being serialized as a `pathlib.Path` instead of a string
when writing to `registered_users.yaml`. The return of the call to
`subprocess.check_output` inside of `get_beatcloud_tracks` was not being
decoded as a string.

What?
Have `USB_PATH` be a string type by default. Add a validator to
`SyncConfig` to convert a non-empty string to a `pathlib.Path`. Add
validation upon `SyncConfig` initialization to ensure `USB_PATH` exists
before performing either the `DOWNLOAD_MUSIC` or `UPLOAD_MUSIC`
operation. Cast `USB_PATH` to a string before writing it to
`registered_users.yaml`. Fix bug where the output of
`subprocess.check_output` was not being decoded.