feat(scripts): analyze Rekordbox My Tags by user
Why?
It's cool to see the distribution of tags by users of my Beatcloud.

What?
Group tracks in an XML by the part of the `Location` path that
corresponds with username, search the `Comments` field for the regex
matching `My Tags` data, counting the frequency of each tag, and then
plot histograms per user.