feat(rekordbox): deprecate registered_users.yaml
Why?
Users shouldn't have to maintain a list of registered users. The whole
point of this file is to facilitate `Location` field manipulations as
part of `rewrite_xml` but, since the `Location` field has a hardcoded
substring of "DJ Music", we can infer all users' `USB_PATH`.

What?
Remove the usage of `registered_users.yaml`.