feat(version): add single source package version
Why?
A `__version__` variable and `--version` option should be available.

What?
Add a `version.py` module containing a `__version__` variable which is:
  - used to setup the package
  - exported from the `__init__` module
  - printed with the `--version` option

test: create namespace fixture

Why?
Initializing a `Namespace` object where ever `argparse` is mocked is too
verbose.

What?
Create a fixture which returns a default `Namespace`.

refactor(configs): improve config print upon execution

Why?
Interpreting the config in the console is difficult.

What?
Override the `__repr__` method of `BaseConfig` to print only the
attributes of the sub-class. Format the representation with new lines
and indentations to improve readability.