script: remove tags and populate fields
Why?
My library is missing `year`, `album`, and `label` tags. I want to
remove "situation" tags ("Opener", "Build", "Peak Time") because I'm
repurposing the `rating` field to convey a track's energy level.

What?
Adds a script to map a threaded function to a collection's tracks. The
function hits the Spotify API to find a matching track and resolves the
`album`, `year`, and `label` fields. In addition, the `comments` are
parsed for "My Tag" data and "situation" tags are removed.