fix: normalize audio bugs
Why?
There are two bugs in the `normalize_audio` module.

The first bug is writing an output file with the same name
(file extension) as the input file when the file format is changed.

For example, if a `wav` is given as input with `AUDIO_FORMAT` set to
`mp3`, the module would export an mp3 file with a `wav` file extension
which causes confusing errors when trying to read the track.

The second bug is an unhandled `CouldntDecodeError` when trying to load
files that aren't audio files.

What?
Instead of writing the output filename as the input filename, build the
output filename using the configured `AUDIO_FORMAT` as the extension.

Filter the list of input tracks for `is_file()` and
`not startswith(".")`. Also wrap the file loading with a `try / except`
block and log an error.