refactor: show full BaseConfig
Why?
When displaying configs during CLI execution, it's desirable to show
each package's config separately with the BaseConfig containing only
the fields exclusive to it so as to reduce clutter and organize the
output...

When displaying the BaseConfig in, for example, a Python shell, it's
much more useful to show the entire config instead having to access
directly the fields which you cannot see.

What?
Inspect the stack frame to infer whether build_config is running as part
of the CLI's execution or not.