feat: override playlist names (#229)
Why?
Combiner playlists are created using expressions that are defined in the
name of the playlists themselves. As a result, playlists can have very
long names. Renaming them after importing is always an option but should
be done automatically when generating the playlist.

What?
Add a PlaylistName class to be used in collection_playlists.yaml.